requirements.txt
psycopg2
python-dotenv
tqdm
pandas